requirements.txt
streamlit
pandas
openpyxl
boto3
Pillow